web/requirements.txt
bert-serving-client==1.10.0
elasticsearch==7.9.1
Flask==1.1.1
